+++ NostalgiaBack/Controllers/TagsController.cs
@@ -19,6 +19,11 @@ namespace NostalgiaBack.Controllers
         {
             return Ok(await _tagsService.GetAllAsync());
         }
+        [HttpGet("{id}")]
+        public async Task<IActionResult> Get([FromRoute] int id)  //default FromQuery   => ~/api/movies?id=2   //FromRoute => ~/api/movies/2
+        {
+            return Ok(await _tagsService.GetByIdAsync(id));
+        }
 
         [HttpPost("Create")]
         public async Task<IActionResult> Create(TagCreateDTO tag)
@@ -27,15 +32,29 @@ namespace NostalgiaBack.Controllers
             return Ok();
         }
 
-        [HttpDelete]
+        [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var existingTag = await _tagsService.GetByIdAsync(id);
+            if (existingTag == null)
+            {
+                return NotFound();
+            }
+
             await _tagsService.DeleteAsync(id);
             return Ok();
         }
+
         [HttpPut("Edit")]
         public async Task<IActionResult> Edit(TagDTO tag)
         {
+            // Check if the tag with the provided id exists
+            var existingTag = await _tagsService.GetByIdAsync(tag.Id);
+            if (existingTag == null)
+            {
+                return NotFound(); // Return NotFound if the tag doesn't exist
+            }
+
             await _tagsService.EditAsync(tag);
             return Ok();
         }
